api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
alembic==1.13.1  # For database migrations
PyMySQL==1.1.0  # Pure Python MySQL client
mysql-connector-python==8.3.0  # For database setup script

# Optional: PostgreSQL support
# psycopg2-binary==2.9.9  # Uncomment if using PostgreSQL
